serverless-mcp/client/requirements.txt
awslabs.mcp_lambda_handler
langchain_aws
langchain_community
langchain_qdrant
pyjokes
python-dotenv
qdrant-client
pypdf
strands-agents 
strands-agents-tools
streamlit
firecrawl-py